ct/api/python/requirements.txt
PyJWT
flask
flask_cors
jsonify
toml
gunicorn